Tighten access controls for AzureStorageBlob
This commit also removes classes that aren't needed in the current
implementation, and moves things around into more logical groups within
the .xcodeproj
