+++ sdk/storage/AzureStorageBlob/Source/DataStructures/StorageBlobClientOptions.swift
