+++ sdk/storage/AzureStorageBlob/Source/StorageHTTPHeader.swift
