+++ sdk/storage/AzureStorageBlob/Source/StorageStreamDownloader.swift
