+++ sdk/storage/AzureStorageBlob/Source/StorageStreamUploader.swift
